Moved record class
