+++ lib/Adapter/Php/InMemory/InMemoryQuery.php
@@ -5,7 +5,7 @@ namespace Phpactor\Indexer\Adapter\Php\InMemory;
 use Phpactor\Indexer\Model\Record\FunctionRecord;
 use Phpactor\Name\FullyQualifiedName;
 use Phpactor\Indexer\Model\IndexQuery;
-use Phpactor\Indexer\Model\Record\ClassRecord;
+use Phpactor\Indexer\Model\Record;
 
 class InMemoryQuery implements IndexQuery
 {
@@ -32,7 +32,7 @@ class InMemoryQuery implements IndexQuery
         }, $class->implementations());
     }
 
-    public function class(FullyQualifiedName $name): ?ClassRecord
+    public function class(FullyQualifiedName $name): ?Record
     {
         return $this->repository->getClass($name->__toString());
     }
