+++ lib/Adapter/Php/Serialized/SerializedQuery.php
@@ -5,7 +5,7 @@ namespace Phpactor\Indexer\Adapter\Php\Serialized;
 use Phpactor\Indexer\Model\Record\FunctionRecord;
 use Phpactor\Name\FullyQualifiedName;
 use Phpactor\Indexer\Model\IndexQuery;
-use Phpactor\Indexer\Model\Record\ClassRecord;
+use Phpactor\Indexer\Model\Record;
 
 class SerializedQuery implements IndexQuery
 {
@@ -35,7 +35,7 @@ class SerializedQuery implements IndexQuery
         }, $class->implementations());
     }
 
-    public function class(FullyQualifiedName $name): ?ClassRecord
+    public function class(FullyQualifiedName $name): ?Record
     {
         return $this->repository->getClass($name);
     }
