+++ lib/Model/IndexQuery.php
@@ -4,7 +4,7 @@ namespace Phpactor\Indexer\Model;
 
 use Phpactor\Indexer\Model\Record\FunctionRecord;
 use Phpactor\Name\FullyQualifiedName;
-use Phpactor\Indexer\Model\Record\ClassRecord;
+use Phpactor\Indexer\Model\Record;
 
 interface IndexQuery
 {
@@ -13,7 +13,7 @@ interface IndexQuery
      */
     public function implementing(FullyQualifiedName $name): array;
 
-    public function class(FullyQualifiedName $name): ?ClassRecord;
+    public function class(FullyQualifiedName $name): ?Record;
 
     public function function(FullyQualifiedName $fullyQualifiedName): ?FunctionRecord;
 }
